Se implemento el CRUD al modulo Categoria y Proveedor asi como un pop up para eliminar en Productos .

+++ src/pages/Products/Productos.jsx
@@ -2,27 +2,30 @@ import Table from "../../components/Table";
 import useStoreProduct from "../../store/useStoreProducts";
 import {
     Pencil,
-    Eraser,
+    Trash,
     Eye,
     CirclePlus
 } from "lucide-react"
+
+import Modal from "../../components/Modal";
 
 import Breadcrumbs from "../../components/Breadcrumbs ";
 
 import { useNavigate } from "react-router-dom";
 
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
 
 import toast from "react-hot-toast";
 import useToastStore from "../../store/toastStore";
-
-import { Tooltip } from "react-tooltip";
 
 export default function Productos() {
 
     const { productos, fetchProducts, loading, error, deleteProduct } = useStoreProduct()
     const { showToast } = useToastStore()
     const navigate = useNavigate()
+
+    const [open, setOpen] = useState(false);
+    const [selectedProductId, setSelectedProductId] = useState(null); // Estado para el ID del producto seleccionado
 
     //! EditProduct
     const handleEdit = (product) => {
@@ -40,25 +43,28 @@ export default function Productos() {
 
     //! DeleteProduct
     const handleDelete = async (productId) => {
-        toast.promise(
-            deleteProduct(productId),
-            {
-                loading: 'Eliminando producto...',
-                success: () => {
-                    // Aquí usamos el store de Zustand para mostrar el toast
-                    showToast('Producto eliminado con éxito!', 'success');
-                    navigate('/productos'); // Redirige a la lista de productos
-                    fetchProducts()
-                    return 'Producto eliminado con éxito!'; // Mensaje de éxito
-                },
-                error: () => {
-                    // También usamos el store de Zustand aquí
-                    showToast('No se pudo eliminar el producto.', 'error');
-                    return 'No se pudo eliminar el producto.'; // Mensaje de error
-                },
+        if (selectedProductId) {
+            toast.promise(
+                deleteProduct(productId),
+                {
+                    loading: 'Eliminando producto...',
+                    success: () => {
+                        // Aquí usamos el store de Zustand para mostrar el toast
+                        showToast('Producto eliminado con éxito!', 'success');
+                        navigate('/productos'); // Redirige a la lista de productos
+                        fetchProducts()
+                        return 'Producto eliminado con éxito!'; // Mensaje de éxito
+                    },
+                    error: () => {
+                        // También usamos el store de Zustand aquí
+                        showToast('No se pudo eliminar el producto.', 'error');
+                        return 'No se pudo eliminar el producto.'; // Mensaje de error
+                    },
 
-            }
-        );
+                }
+            );
+            setOpen(false);
+        }
     }
 
     //! FetchProducts
@@ -84,7 +90,7 @@ export default function Productos() {
             accessorKey: "estadoProducto",
             cell: ({ row }) => {
                 const estado = row.original.estadoProducto
-                const estadoClass = estado === "Alta" ? "bg-green-300 text-green-600" : "bg-red-300 text-red-600"; // Cambia a los colores que desees
+                const estadoClass = estado === "Alta" ? "bg-emerald-100/60 text-emerald-500" : "bg-red-100/60 text-red-500"; // Cambia a los colores que desees
                 return <span className={`font-semibold px-2 py-0 rounded-lg items-center ${estadoClass}`}>{estado}</span>;
             }
         },
@@ -93,12 +99,17 @@ export default function Productos() {
             accessorKey: "acciones",
             cell: ({ row }) => (
                 <div className="flex items-center space-x-2">
-                    <button onClick={() => handleEdit(row.original)} data-tooltip-id="editTooltip" data-tooltip-content="Editar" className="px-2 py-1 text-white bg-indigo-400 rounded hover:bg-indigo-300"><Pencil size={20} strokeWidth={2.5} /></button>
-                    <button onClick={() => handleDelete(row.original.productoId)} data-tooltip-id="deleteTooltip" data-tooltip-content="Eliminar" className="px-2 py-1 text-white bg-red-400 rounded hover:bg-red-300"><Eraser size={20} strokeWidth={2.5} /></button>
-                    <button onClick={() => handleView(row.original)} data-tooltip-id="viewTooltip" data-tooltip-content="Ver" className="px-2 py-1 text-white bg-green-400 rounded hover:bg-green-300"><Eye size={20} strokeWidth={2.25} /></button>
-                    <Tooltip id="editTooltip" place="top" />
-                    <Tooltip id="deleteTooltip" place="top" />
-                    <Tooltip id="viewTooltip" place="top" />
+                    <button onClick={() => handleEdit(row.original)} className="px-2 py-1 text-indigo-500"><Pencil size={20} strokeWidth={2.5} /></button>
+                    <button
+                        onClick={() => {
+                            setSelectedProductId(row.original.productoId);
+                            setOpen(true);
+                        }}
+                        className="px-2 py-1 text-red-500"
+                    >
+                        <Trash size={20} strokeWidth={2.25} />
+                    </button>
+                    <button onClick={() => handleView(row.original)} className="px-2 py-1 text-emerald-500"><Eye size={20} strokeWidth={2.25} /></button>
                 </div>
             )
         }
@@ -114,7 +125,7 @@ export default function Productos() {
             />
 
             <Table
-                products={productos}
+                data={productos}
                 columns={columns}
                 fetchProducts={fetchProducts}
                 loading={loading}
@@ -129,6 +140,34 @@ export default function Productos() {
                     subtitle: "Lista de todos los productos."
                 }}
             />
+
+            {/* Modal */}
+            <Modal open={open} onClose={() => setOpen(false)}>
+                <div className="text-center w-56">
+                    <Trash size={56} className="mx-auto text-red-500" />
+                    <div className="mx-auto my-4 w-48">
+                        <h3 className="text-lg font-black text-gray-800">Confirmar Eliminacion</h3>
+                        <p className="text-sm text-gray-500">
+                            Estas seguro de eliminar este producto?
+                        </p>
+                    </div>
+                    <div className="flex gap-4">
+                        <button
+                            className="w-full p-2 bg-red-100/60 text-red-500 rounded hover:bg-red-200 font-semibold transition duration-200"
+                            onClick={() => handleDelete(selectedProductId)}
+                        >
+                            Delete
+                        </button>
+
+                        <button
+                            className="w-full p-2 bg-emerald-100/60 text-emerald-500 rounded hover:bg-emerald-200 font-semibold transition duration-200"
+                            onClick={() => setOpen(false)}
+                        >
+                            Cancel
+                        </button>
+                    </div>
+                </div>
+            </Modal>
         </div>
     )
 }
